+++ packages/components/tabs/src/index.ts
@@ -1,0 +1,12 @@
+import Tabs from "./tabs";
+
+// export types
+export type {TabsProps} from "./tabs";
+
+// export hooks
+export {useTabs} from "./use-tabs";
+
+// export component
+export {Tabs};
+
+export {default as Tab} from "./base/tab-item-base";