+++ packages/core/theme/src/components/index.ts
@@ -26,7 +26,7 @@ export * from "./image";
 export * from "./spacer";
 export * from "./divider";
 export * from "./kbd";
-// export * from "./tabs";
+export * from "./tabs";
 export * from "./skeleton";
 // export * from "./select";
 export * from "./menu";