+++ packages/components/link/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jala-banyu/link
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.3.0
+
 ## 3.0.0
 
 ### Patch Changes