+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/popover
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.3.0
+  - @jala-banyu/button@4.0.0
+
 ## 2.0.0
 
 ### Patch Changes